refactor(web): move keyboard shortcut to design system
