+++ web/src/components/design-system/KeyboardShortcut/KeyboardShortcut.stories.tsx
@@ -1,4 +1,4 @@
-import preview from "../../../.storybook/preview";
+import preview from "../../../../.storybook/preview";
 import { KeyboardShortcut } from "./keyboard-shortcut";
 
 const meta = preview.meta({
