+++ web/src/components/design-system/KeyboardShortcut/keyboard-shortcut.tsx
