+++ web/src/features/evals/components/code-eval-template-form-body.tsx
@@ -24,7 +24,7 @@ import {
 import { Loader2 } from "lucide-react";
 
 import { Button } from "@/src/components/ui/button";
-import { KeyboardShortcut } from "@/src/components/ui/keyboard-shortcut";
+import { KeyboardShortcut } from "@/src/components/design-system/KeyboardShortcut/keyboard-shortcut";
 import {
   Tooltip,
   TooltipContent,
